chore: change mixed_precision to bf16

+++ verl/workers/fsdp_workers.py
@@ -186,8 +186,8 @@ class ActorRolloutRefWorker(Worker):
             buffer_dtype = PrecisionType.to_dtype(mixed_precision_config.get('buffer_dtype', 'fp32'))
         else:
             param_dtype = torch.bfloat16
-            reduce_dtype = torch.float32
-            buffer_dtype = torch.float32
+            reduce_dtype = torch.bfloat16
+            buffer_dtype = torch.bfloat16
 
         mixed_precision = MixedPrecision(param_dtype=param_dtype, reduce_dtype=reduce_dtype, buffer_dtype=buffer_dtype)
